src/mcp-server/03-banking-server/requirements.txt
agent-framework
openai
python-dotenv
uvicorn
fastapi
mcp-server
